// generated/src/app/api-models/incentive-quality-pgm.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IncentiveQualityPgm {
  seqQualityPgm: number;
  seqIncentiveRule: number;
  qualityProgram: string;
  qualityIncentiveAmount: number;
  qualityCapApply: string;
  qualityIncentivePercent: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}